ersatz_rag/regulus/backend/pyproject.toml
[project]
name = "regulus-backend"
version = "0.1.0"
description = "REGULUS Enterprise Policy & Compliance Platform Backend"
readme = "README_SIMPLE.md"
requires-python = ">=3.12"
dependencies = [
    # Core Framework
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",

    # Database & Storage
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "redis>=5.0.0",
    "arq>=0.25.0",

    # AI & ML
    "openai>=1.0.0",
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",

    # ERSATZ_RAG Components
    "leann-core",
    "leann-backend-hnsw",
    "pageindex",

    # Search & Vector Stores
    "qdrant-client>=1.6.0",
    "opensearch-py>=2.3.0",
    "faiss-cpu>=1.7.0",

    # Document Processing
    "PyMuPDF>=1.23.0",
    "python-multipart>=0.0.6",

    # HTTP & Requests
    "requests>=2.31.0",
    "httpx>=0.25.0",

    # CLI & Utils
    "typer>=0.9.0",
    "rich>=13.0.0",
    "tqdm>=4.66.0",

    # Development & Testing
    "pytest>=7.4.0",
    "ruff>=0.1.0"
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "C90", "I", "N", "UP", "B", "A", "C4", "EM", "FA", "ISC", "T20", "Q"]
ignore = ["E501", "UP007"]
